backend/mqtt_forwarder/requirements.txt
certifi==2025.8.3
click==8.3.0
Flask==2.0.3
influxdb-client==1.31.0
influxdb3-python==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
paho-mqtt==2.1.0
pyarrow==21.0.0
python-dateutil==2.9.0.post0
reactivex==4.0.4
Rx==3.2.0
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==2.2.2
